🔈<Log>: #21 로그인,회원가입 기능 업데이트
로그인과 회원가입 할 때 이메일 인증 방식을 분리해서 구현했고 api 주소를 다르게 설정해 오류가 생기는걸 방지함
프로젝트의 group과 name을 변경함
이메일 인증 요청을 보낼 때 중간에 누군가 같은 이메일로 회원가입을 했을 경우를 대비해 항상 이메일 존재 여부를 검사하게 수정
