+++ server/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'goose'
+rootProject.name = 'Kok'
