--- server/src/main/java/com/example/goose/iGoose/auth/request/LoginRequest.java
@@ -1,16 +1,0 @@
-package com.example.goose.iGoose.auth.request;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginRequest {
-    private String id;
-    private String password;
-}
