--- server/src/main/java/com/example/goose/iGoose/auth/vo/VerificationVO.java
@@ -1,24 +1,0 @@
-package com.example.goose.iGoose.auth.vo;
-
-
-import lombok.*;
-
-import java.util.Date;
-
-@Getter
-@Setter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-public class VerificationVO {
-
-    private String uuid;
-
-    private String method;
-
-    private String code;
-
-    private Date expire;
-
-    private Boolean is_verified;
-}
