+++ server/src/main/java/com/iGoose/Kok/common/config/SecurityConfig.java
@@ -1,0 +1,67 @@
+package com.iGoose.Kok.common.config;
+
+import com.iGoose.Kok.common.filter.JwtTokenFilter;
+import com.iGoose.Kok.common.service.CustomUserDetailsService;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import javax.sql.DataSource;
+
+@Configuration
+@EnableWebSecurity
+@RequiredArgsConstructor
+public class SecurityConfig {
+
+    private final JwtTokenFilter jwtTokenFilter;
+    private final CustomUserDetailsService userDetailsService;
+    private DataSource dataSource;
+
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http
+                .csrf(csrf -> csrf.disable())  // CSRF 보호 비활성화
+                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))  // 세션 관리
+                .authorizeHttpRequests(auth -> auth
+                        .requestMatchers("/auth/**").permitAll()  // 인증 없이 접근 허용
+                        .anyRequest().authenticated()  // 나머지 요청은 인증 필요
+                )
+                .rememberMe(rememberMe -> rememberMe
+                        .rememberMeParameter("remember")
+                        .tokenValiditySeconds(3600)
+                        .userDetailsService(userDetailsService))
+                .logout(logout -> logout
+                        .permitAll()
+                        .logoutSuccessUrl("/auth/login")
+                        .deleteCookies("JSESSIONID", "remember"))
+                .addFilterBefore(jwtTokenFilter, UsernamePasswordAuthenticationFilter.class);  // JWT 필터 추가
+
+
+        return http.build();
+    }
+
+    // 비밀번호 암호화를 위한 PasswordEncoder 빈 등록
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    // AuthenticationManager 빈 등록
+    @Bean
+    public AuthenticationEntryPoint authenticationEntryPoint() {
+        return (request, response, authException) -> response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Unauthorized");
+    }
+
+
+}
+
