+++ server/src/main/java/com/iGoose/Kok/iGoose/auth/request/LoginRequest.java
@@ -1,0 +1,16 @@
+package com.iGoose.Kok.iGoose.auth.request;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginRequest {
+    private String id;
+    private String password;
+}
