+++ server/src/main/java/com/iGoose/Kok/iGoose/auth/response/VerificationResponse.java
@@ -1,0 +1,19 @@
+package com.iGoose.Kok.iGoose.auth.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.Date;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class VerificationResponse {
+    private String uuid;
+    private String method;
+    private Date expire;
+    private String code;
+}
