+++ server/src/main/java/com/iGoose/Kok/iGoose/auth/response/VerificationUserInfoResponse.java
@@ -1,0 +1,16 @@
+package com.iGoose.Kok.iGoose.auth.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class VerificationUserInfoResponse {
+    private String message;
+    private String user_name;
+    private String profile;
+}
