+++ server/src/main/java/com/iGoose/Kok/iGoose/auth/vo/VerificationVO.java
@@ -1,0 +1,22 @@
+package com.iGoose.Kok.iGoose.auth.vo;
+
+
+import lombok.*;
+
+import java.util.Date;
+
+@Getter
+@Setter
+@ToString
+@AllArgsConstructor
+@NoArgsConstructor
+public class VerificationVO {
+
+    private String uuid;
+
+    private String method;
+
+    private String code;
+
+    private Date expire;
+}
