--- server/src/main/resources/mapper/auth-SQL.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.goose.iGoose.auth.mapper.AuthMapper">
-
-    <insert id="insertUser" parameterType="UserVO">
-        INSERT INTO "USER" (uuid,
-                            id,
-                            password,
-                            created,
-                            updated)
-        VALUES (#{uuid},
-                #{id},
-                #{password},
-                now(),
-                now());
-    </insert>
-
-</mapper>
