--- server/src/test/java/com/example/goose/GooseApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.goose;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GooseApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
