+++ server/src/test/java/com/iGoose/Kok/GooseApplicationTests.java
@@ -1,0 +1,13 @@
+package com.iGoose.Kok;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GooseApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
